+++ detekt-api/src/test/kotlin/io/gitlab/arturbosch/detekt/api/ConfigSpec.kt
@@ -1,12 +1,12 @@
 package io.gitlab.arturbosch.detekt.api
 
+import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions.assertThatExceptionOfType
+import org.assertj.core.api.Assertions.fail
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
 import java.nio.file.Paths
-import kotlin.test.assertFailsWith
-import kotlin.test.assertTrue
-import kotlin.test.fail
 
 /**
  * @author Artur Bosch
@@ -20,11 +20,11 @@ class ConfigSpec : Spek({
 		it("should create a sub config") {
 			try {
 				val subConfig = config.subConfig("style")
-				assertTrue { subConfig.valueOrDefault("WildcardImport", mapOf<String, Any>()).isNotEmpty() }
-				assertTrue { subConfig.valueOrDefault("WildcardImport", mapOf<String, Any>())["active"].toString() == "true" }
-				assertTrue { subConfig.valueOrDefault("WildcardImport", mapOf<String, Any>())["active"] as Boolean }
-				assertTrue { subConfig.valueOrDefault("NotFound", mapOf<String, Any>()).isEmpty() }
-				assertTrue { subConfig.valueOrDefault("NotFound", "") == "" }
+				assertThat(subConfig.valueOrDefault("WildcardImport", mapOf<String, Any>())).isNotEmpty
+				assertThat(subConfig.valueOrDefault("WildcardImport", mapOf<String, Any>())["active"].toString()).isEqualTo("true")
+				assertThat(subConfig.valueOrDefault("WildcardImport", mapOf<String, Any>())["active"] as Boolean).isTrue()
+				assertThat(subConfig.valueOrDefault("NotFound", mapOf<String, Any>())).isEmpty()
+				assertThat(subConfig.valueOrDefault("NotFound", "")).isEmpty()
 			} catch (ignored: Config.InvalidConfigurationError) {
 				fail("Creating a sub config should work for test resources config!")
 			}
@@ -34,15 +34,15 @@ class ConfigSpec : Spek({
 			try {
 				val subConfig = config.subConfig("style")
 				val subSubConfig = subConfig.subConfig("WildcardImport")
-				assertTrue { subSubConfig.valueOrDefault("active", false) }
-				assertTrue { subSubConfig.valueOrDefault("NotFound", true) }
+				assertThat(subSubConfig.valueOrDefault("active", false)).isTrue()
+				assertThat(subSubConfig.valueOrDefault("NotFound", true)).isTrue()
 			} catch (ignored: Config.InvalidConfigurationError) {
 				fail("Creating a sub config should work for test resources config!")
 			}
 		}
 
 		it("tests wrong sub config conversion") {
-			assertFailsWith<ClassCastException> {
+			assertThatExceptionOfType(ClassCastException::class.java).isThrownBy {
 				@Suppress("UNUSED_VARIABLE")
 				val ignored = config.valueOrDefault("style", "")
 			}