+++ detekt-api/src/test/kotlin/io/gitlab/arturbosch/detekt/api/MetricSpec.kt
@@ -1,9 +1,9 @@
 package io.gitlab.arturbosch.detekt.api
 
+import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions.assertThatIllegalStateException
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.it
-import kotlin.test.assertEquals
-import kotlin.test.assertFails
 
 /**
  * @author Artur Bosch
@@ -12,12 +12,12 @@ class MetricSpec : Spek({
 
 	it("should convert double values to int") {
 		val metric = Metric("LOC", 0.33, 0.10, 100)
-		assertEquals(0.33, metric.doubleValue())
-		assertEquals(0.10, metric.doubleThreshold())
+		assertThat(metric.doubleValue()).isEqualTo(0.33)
+		assertThat(metric.doubleThreshold()).isEqualTo(0.10)
 	}
 
 	it("should throw error if double value is asked for int metric") {
-		assertFails {
+		assertThatIllegalStateException().isThrownBy {
 			Metric("LOC", 100, 50).doubleValue()
 		}
 	}