+++ detekt-gradle-plugin/src/test/kotlin/io/gitlab/arturbosch/detekt/DetektTaskKotlinDslTest.kt
@@ -13,6 +13,21 @@ import java.io.File
  */
 internal class DetektTaskKotlinDslTest : Spek({
 
+	val buildGradle = """
+		|import io.gitlab.arturbosch.detekt.detekt
+		|
+		|plugins {
+		|   `java-library`
+		|	id("io.gitlab.arturbosch.detekt")
+		|}
+		|
+		|repositories {
+		|	jcenter()
+		|	mavenLocal()
+		|}
+		""".trimMargin()
+	val dslTest = DslBaseTest("build.gradle.kts", buildGradle)
+
 	describe("The Detekt Gradle plugin used in a build.gradle.kts file") {
 		lateinit var rootDir: File
 		beforeEachTest {
@@ -22,8 +37,8 @@ internal class DetektTaskKotlinDslTest : Spek({
 
 			val detektConfig = ""
 
-			writeFiles(rootDir, detektConfig)
-			writeConfig(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig)
+			dslTest.writeConfig(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -41,8 +56,9 @@ internal class DetektTaskKotlinDslTest : Spek({
 
 			val detektConfig = ""
 
-			writeFiles(rootDir, detektConfig, "src/main/kotlin")
-			writeConfig(rootDir)
+			val srcDir = File(rootDir, "src/main/kotlin")
+			dslTest.writeFiles(rootDir, detektConfig, srcDir)
+			dslTest.writeConfig(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -64,8 +80,8 @@ internal class DetektTaskKotlinDslTest : Spek({
 					|}
 				"""
 
-			writeFiles(rootDir, detektConfig)
-			writeConfig(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig)
+			dslTest.writeConfig(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -84,15 +100,16 @@ internal class DetektTaskKotlinDslTest : Spek({
 		}
 
 		it("can be applied with a custom config file") {
+			val configPath = File(rootDir, "config.yml")
 
 			val detektConfig = """
 					|detekt {
-					|	config = files("${rootDir.absolutePath}/config.yml")
+					|	config = files("${configPath.safeAbsolutePath}")
 					|}
 				"""
 
-			writeFiles(rootDir, detektConfig)
-			writeConfig(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig)
+			dslTest.writeConfig(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -101,12 +118,13 @@ internal class DetektTaskKotlinDslTest : Spek({
 					.withPluginClasspath()
 					.build()
 
-			assertThat(result.output).contains("Ruleset: comments", "number of classes: 1", "--config ${rootDir.absolutePath}/config.yml")
+			assertThat(result.output).contains("Ruleset: comments", "number of classes: 1", "--config ${configPath.absolutePath}")
 			assertThat(result.task(":check")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
 		}
 
 		it("can be applied with a full config") {
-
+			val configFile = File(rootDir, "config.yml")
+			val baselineFile = File(rootDir, "baseline.xml")
 
 			val detektConfig = """
 					|detekt {
@@ -114,15 +132,15 @@ internal class DetektTaskKotlinDslTest : Spek({
 					|	parallel = true
 					|	disableDefaultRuleSets = true
 					|	toolVersion = "$VERSION_UNDER_TEST"
-					|	config = files("${rootDir.absolutePath}/config.yml")
-					|	baseline = file("${rootDir.absolutePath}/baseline.xml")
+					|	config = files("${configFile.safeAbsolutePath}")
+					|	baseline = file("${baselineFile.safeAbsolutePath}")
 					|	filters = ".*/resources/.*, .*/build/.*"
 					|}
 				"""
 
-			writeFiles(rootDir, detektConfig)
-			writeConfig(rootDir)
-			writeBaseline(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig)
+			dslTest.writeConfig(rootDir)
+			dslTest.writeBaseline(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -138,13 +156,14 @@ internal class DetektTaskKotlinDslTest : Spek({
 		}
 
 		it("can configure a new custom detekt task") {
+			val configFile = File(rootDir, "config.yml")
 
 			val detektConfig = """
 					|task<io.gitlab.arturbosch.detekt.Detekt>("detektFailFast") {
 					|	description = "Runs a failfast detekt build."
 					|
 					|	input = files("src/main/java")
-					|	config = files("$rootDir/config.yml")
+					|	config = files("${configFile.safeAbsolutePath}")
 					|	debug = true
 					|	reports {
 					|		xml {
@@ -155,9 +174,9 @@ internal class DetektTaskKotlinDslTest : Spek({
 					|}
 				"""
 
-			writeFiles(rootDir, detektConfig)
-			writeConfig(rootDir, failfast = true)
-			writeBaseline(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig)
+			dslTest.writeConfig(rootDir, failfast = true)
+			dslTest.writeBaseline(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -185,9 +204,9 @@ internal class DetektTaskKotlinDslTest : Spek({
 					|}
 				"""
 
-			writeFiles(rootDir, detektConfig)
-			writeConfig(rootDir)
-			writeBaseline(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig)
+			dslTest.writeConfig(rootDir)
+			dslTest.writeBaseline(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -211,9 +230,9 @@ internal class DetektTaskKotlinDslTest : Spek({
 					|}
 				"""
 
-			writeFiles(rootDir, detektConfig)
-			writeConfig(rootDir)
-			writeBaseline(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig)
+			dslTest.writeConfig(rootDir)
+			dslTest.writeBaseline(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -230,17 +249,17 @@ internal class DetektTaskKotlinDslTest : Spek({
 		}
 
 		it("can configure the input directory") {
-			val customSourceLocation = "gensrc/kotlin"
+			val customSourceLocation = File(rootDir, "gensrc/kotlin")
 			val detektConfig = """
 					|detekt {
 					| debug = true
-					| input = files("$customSourceLocation")
+					| input = files("${customSourceLocation.safeAbsolutePath}")
 					|}
 				"""
 
-			writeFiles(rootDir, detektConfig, customSourceLocation)
-			writeConfig(rootDir)
-			writeBaseline(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig, customSourceLocation)
+			dslTest.writeConfig(rootDir)
+			dslTest.writeBaseline(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -250,19 +269,19 @@ internal class DetektTaskKotlinDslTest : Spek({
 					.build()
 
 			assertThat(result.output).contains("number of classes: 1")
-			assertThat(result.output).contains("--input, $rootDir/$customSourceLocation")
+			assertThat(result.output).contains("--input, ${customSourceLocation.absolutePath}")
 			assertThat(result.task(":check")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
 		}
 
 		it("can be used without configuration") {
 			val detektConfig = """
 					|dependencies {
-					| detekt("io.gitlab.arturbosch.detekt:detekt-formatting:${VERSION_UNDER_TEST}")
+					| detektPlugins("io.gitlab.arturbosch.detekt:detekt-formatting:${VERSION_UNDER_TEST}")
 					|}
 				"""
 
-			writeFiles(rootDir, detektConfig)
-			writeConfig(rootDir)
+			dslTest.writeFiles(rootDir, detektConfig)
+			dslTest.writeConfig(rootDir)
 
 			// Using a custom "project-cache-dir" to avoid a Gradle error on Windows
 			val result = GradleRunner.create()
@@ -278,47 +297,3 @@ internal class DetektTaskKotlinDslTest : Spek({
 	}
 })
 
-// build.gradle.kts
-private fun getBuildFileContent(detektConfig: String) = """
-	|import io.gitlab.arturbosch.detekt.detekt
-	|
-	|plugins {
-	|   `java-library`
-	|	id("io.gitlab.arturbosch.detekt")
-	|}
-	|
-	|repositories {
-	|	jcenter()
-	|	mavenLocal()
-	|}
-	|
-	|$detektConfig
-	""".trimMargin()
-
-
-// src/main/kotlin/MyClass.kt
-private val ktFileContent = """
-	|internal class MyClass
-	|
-	""".trimMargin()
-
-private fun writeFiles(root: File, detektConfig: String, sourceDir: String = "src/main/java") {
-	File(root, "build.gradle.kts").writeText(getBuildFileContent(detektConfig))
-	File(root, sourceDir).mkdirs()
-	File(root, "$sourceDir/MyClass.kt").writeText(ktFileContent)
-}
-
-private fun writeConfig(root: File, failfast: Boolean = false) {
-	File(root, "config.yml").writeText("""
-		|autoCorrect: true
-		|failFast: $failfast
-		""".trimMargin())
-}
-
-private fun writeBaseline(root: File) {
-	File(root, "baseline.xml").writeText("""
-		|<some>
-		|	<xml/>
-		|</some>
-		""".trimMargin())
-}